% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_censored_df}
\alias{add_censored_df}
\title{Consruct fitdistcens-compatible input from {{ time }} and {{ fail }} columns}
\usage{
add_censored_df(data, time = time, fail = fail)
}
\arguments{
\item{data}{A data frame of (possibly) censored lifetimes}

\item{time}{A quasiquoted column of lifetimes}

\item{fail}{A quasiquoted column of censoring info}
}
\value{
A data frame appended with left/right
}
\description{
Consruct fitdistcens-compatible input from {{ time }} and {{ fail }} columns
}
\examples{
lifetimes <- data.frame(
  time = 1:5,
  cnsrd = c(TRUE, rep(FALSE, 4)))

add_censored_df(lifetimes, fail = !cnsrd)
}
