% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_km}
\alias{add_km}
\title{Append(?) km points for probability plotting}
\usage{
add_km(data)
}
\arguments{
\item{data}{A tibble with time and fail columns}
}
\value{
A tibble with km's appended
}
\description{
Append(?) km points for probability plotting
}
