% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_mle_qtile}
\alias{add_mle_qtile}
\title{Append predicted quantiles from survreg model}
\usage{
add_mle_qtile(data, model)
}
\arguments{
\item{data}{A tibble containing p's from which to make predict quantiles}

\item{model}{A survreg model}
}
\value{
A tibble appended with predicted quantiles
}
\description{
Append predicted quantiles from survreg model
}
\details{
p's are gotten from km or ppts
}
