% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_ppts}
\alias{add_ppts}
\title{Append ppts to lifetime tibble}
\usage{
add_ppts(data)
}
\arguments{
\item{data}{A tibble with a time column}
}
\value{
A tibble with 1 - ppts appended
}
\description{
Append ppts to lifetime tibble
}
