% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_surv}
\alias{fit_surv}
\title{Fit a survreg model}
\usage{
fit_surv(data)
}
\arguments{
\item{data}{tibble containing lifetimes and censoring info}
}
\value{
A survreg object
}
\description{
Fit a survreg model
}
