% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_alpha}
\alias{get_alpha}
\title{Extract Weibull alpha from a (tidied) survreg model}
\usage{
get_alpha(tidy)
}
\arguments{
\item{tidy}{a tidied survreg model}
}
\value{
numeric(1) Weibull alpha
}
\description{
Extract Weibull alpha from a (tidied) survreg model
}
