% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_beta}
\alias{get_beta}
\title{Extract Weibull beta from a (tidied) survreg model}
\usage{
get_beta(tidy)
}
\arguments{
\item{tidy}{a tidied survreg model}
}
\value{
numeric(1) Weibull beta
}
\description{
Extract Weibull beta from a (tidied) survreg model
}
