% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_fail_time}
\alias{get_fail_time}
\title{Get failure time by log-linear interpolation}
\usage{
get_fail_time(data, y, time, fail_crit)
}
\arguments{
\item{data}{tibble with value and hrs columns}

\item{y}{quasiquoted response (ie, degradation metric)}

\item{time}{quasiquoted Time under stress}

\item{fail_crit}{Failure criterion (eg, cap change)}
}
\value{
numeric interpolated failure time
}
\description{
Get failure time by log-linear interpolation
}
