% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ppts}
\alias{get_ppts}
\title{Get ppoints (alternative to km)}
\usage{
get_ppts(time)
}
\arguments{
\item{time}{a numeric vector}
}
\value{
numeric vector of plotting positions (the same length as time)
}
\description{
Get ppoints (alternative to km)
}
