% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_stub}
\alias{get_stub}
\title{Get stub from capacitance at two frequencies}
\usage{
get_stub(data)
}
\arguments{
\item{data}{tibble with cap (uF) and frq (kHz)}
}
\value{
numeric(1) capacitance difference
}
\description{
Get stub from capacitance at two frequencies
}
\details{
Stubs are associated with end-spray separation from film end, and hence tend to start after a time "threshold"

Not a stub if max(frq) < 3 kHz, so return 0
}
