% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/components.R
\name{node}
\alias{node}
\title{A series of nodes}
\usage{
node(graph = NULL, name, reliability, n = 1)
}
\arguments{
\item{name}{}

\item{item}{}

\item{exp_time_to_failure}{}
}
\value{
list
}
\description{
A series of nodes
}

