% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/components.R
\name{nodes}
\alias{nodes}
\title{A series of nodes}
\usage{
nodes(system = NULL, name, reliability, n)
}
\arguments{
\item{name}{}

\item{item}{}

\item{exp_time_to_failure}{}
}
\value{
list
}
\description{
A series of nodes
}

