% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/components.R
\name{parallel}
\alias{parallel}
\title{Parallel}
\usage{
parallel(system = NULL, name, components, n)
}
\arguments{
\item{system}{}

\item{name}{}

\item{components}{}

\item{n}{}
}
\value{
list
}
\description{
Parallel
}

