% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_test2.R
\name{Reliability_test2}
\alias{Reliability_test2}
\title{Reliability_test2() Function}
\usage{
Reliability_test2(a, b, c)
}
\arguments{
\item{a, b, c}{:a is data b is numeric}
}
\description{
신뢰성학회 연습용 입니다.
}
\keyword{Reliability_test2}
