% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{weibull_trans}
\alias{weibull_trans}
\title{Probability-model linearizing transformations}
\usage{
weibull_trans()
}
\value{
A scale transformer
}
\description{
Probability-model linearizing transformations
}
\examples{
weibull_trans()
}
