% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/components.R
\name{nodes}
\alias{nodes}
\title{Node}
\usage{
nodes(item = NULL, name, reliability, n = 1)
}
\arguments{
\item{item}{}

\item{name}{}

\item{exp_time_to_failure}{}
}
\value{
list
}
\description{
Node
}

