% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_test.R
\name{Reliability_test}
\alias{Reliability_test}
\title{Reliability_test() Function}
\usage{
Reliability_test(x, t)
}
\arguments{
\item{x, t}{:x is data t is numeric}
}
\description{
신뢰성학회 연습용 입니다.
}
\keyword{Reliability_test}
