% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reliability_test2.R
\name{Reliability_test2}
\alias{Reliability_test2}
\title{Reliability_test2() Function}
\usage{
Reliability_test2(r, Cr, Ci)
}
\arguments{
\item{r:평균고장횟수}{Cr: 고장시 수리비용 Ci:점검비용}
}
\description{
비용최적화를 위한 R 패키지 함수이며 (r,Cr,Ci)형태의 함수를 유지하고 있습니다
r에는 단위시간당 평균고장횟수를 입력 c_r에는 고장시수리비용 c_i에는 점검비용을 입력합니다.
따라서 이를 통해 최적 점검횟수와 총 기대비용이 구해지며 그래프를 통해 기대비용이 어떠한 분포를 이루는지 알 수 있습니다.
}
\keyword{Reliability_test2}
